feat: update project v3 alpha

+++ muttley/src/main/java/com/project/muttley/domain/event/dto/EventSummaryDTO.java
@@ -11,5 +11,5 @@ public record EventSummaryDTO(
     String statusName,
     String typeName,
     String modalityName,
-    Long points) {
+    Integer points) {
 }
